requirements.txt
Django==1.11.3
django-allauth==0.32.0
django-cleanup==1.0.0
django-suit==0.2.25
django-wysiwyg-redactor==0.5.1
Pillow==4.2.1
requests==2.18.1
